+++ R/S3_print.R
@@ -3,6 +3,8 @@ print.lm_robust <- function(x, ...) {
   print(summarize_tidy(x))
 }
 
+
+
 #' @export
 print.iv_robust <- function(x, ...) {
   print(summarize_tidy(x))
@@ -10,6 +12,7 @@ print.iv_robust <- function(x, ...) {
 
 print_summary_lm_like <- function(x,
                                   digits,
+                                  signif.stars = getOption("show.signif.stars"),
                                   ...) {
   cat(
     "\nCall:\n",
@@ -36,7 +39,7 @@ print_summary_lm_like <- function(x,
 
   print(coef(x), digits = digits)
 
-  fstat <- if (is.numeric(x$fstatistic)) {
+  fstat <- if (is.numeric(x[["fstatistic"]])) {
     paste(
       "\nF-statistic:", formatC(x$fstatistic[1L], digits = digits),
       "on", x$fstatistic[2L], "and", x$fstatistic[3L],
@@ -56,7 +59,7 @@ print_summary_lm_like <- function(x,
     fstat
   )
 
-  if (!is.null(x$proj_fstatistic)) {
+  if (is.numeric(x[["proj_fstatistic"]])) {
     cat(
       "\nMultiple R-squared (proj. model): ",
       formatC(x$proj_r.squared, digits = digits),
@@ -76,12 +79,27 @@ print_summary_lm_like <- function(x,
   }
   cat("\n")
 
+  if (is.numeric(x[["diagnostic_endogeneity_test"]])) {
+    cat("\nDiagnostics:\n")
+    printCoefmat(
+      build_ivreg_diagnostics_mat(x),
+      cs.ind = 1L:2L,
+      tst.ind = 3L,
+      has.Pvalue = TRUE,
+      P.values = TRUE,
+      digits = digits,
+      signif.stars = signif.stars,
+      na.print = "NA",
+      ...
+    )
+  }
   invisible(x)
 }
 
 #' @export
 print.summary.lm_robust <- function(x,
                                     digits = max(3L, getOption("digits") - 3L),
+                                    signif.stars = getOption("show.signif.stars"),
                                     ...) {
   print_summary_lm_like(x, digits, ...)
 }
@@ -89,8 +107,9 @@ print.summary.lm_robust <- function(x,
 #' @export
 print.summary.iv_robust <- function(x,
                                     digits = max(3L, getOption("digits") - 3L),
+                                    signif.stars = getOption("show.signif.stars"),
                                     ...) {
-  print_summary_lm_like(x, digits, ...)
+  print_summary_lm_like(x, digits, signif.stars, ...)
 }
 
 #' @export
@@ -103,4 +122,41 @@ print.difference_in_means <- function(x, ...) {
 #' @export
 print.horvitz_thompson <- function(x, ...) {
   print(summarize_tidy(x))
+}
+
+#' @export
+print.lh <- function(x, ...) {
+
+  print(summarize_tidy(x))
+}
+
+#' @export
+print.lh_robust <- function(x, ...) {
+  lnames <- names(x)
+  for (i in seq_along(x)) {
+    cat("$", lnames[i], "\n", sep = "")
+    print(x[[i]])
+    cat("\n")
+  }
+  invisible(x)
+}
+
+
+
+
+#' @export
+print.summary.lh_robust <- function(x,...){
+  cat("$lm_robust \n ")
+  print(summary(x$lm_robust))
+  x <- x[[2]]
+
+cat("\n\n$lh \n \n")
+
+print(attr(x, "linear_hypothesis"))
+}
+#' @export
+print.summary.lh <- function(x, ...){
+
+  print(attr(x, "linear_hypothesis"))
+
 }