+++ R/estimatr_lm_lin.R
@@ -15,9 +15,9 @@
 #' of observations to be used.
 #' @param clusters An optional bare (unquoted) name of the variable that
 #' corresponds to the clusters in the data.
-#' @param se_type The sort of standard error sought. If `clusters` is
+#' @param se_type The sort of standard error sought. If \code{clusters} is
 #' not specified the options are "HC0", "HC1" (or "stata", the equivalent),
-#' "HC2" (default), "HC3", or  "classical". If `clusters` is specified the
+#' "HC2" (default), "HC3", or  "classical". If \code{clusters} is specified the
 #' options are "CR0", "CR2" (default), or "stata" are
 #' permissible.
 #' @param ci logical. Whether to compute and return p-values and confidence
@@ -202,13 +202,10 @@ lm_lin <- function(formula,
   # Get design matrix including `covariates` for centering
   # ----------
 
-  full_formula <-
-    update(
-      formula,
-      reformulate(
-        c(".", labels(cov_terms), response = ".")
-      )
-    )
+  full_formula <- update(
+    formula,
+    reformulate(c(".", labels(cov_terms)))
+  )
 
   datargs <- enquos(
     formula = full_formula,