+++ tests/testthat/test-s3-methods.R
@@ -10,7 +10,7 @@ dat <- data.frame(
 lmbo <- lm_robust(y ~ z + as.factor(x), data = dat)
 lmfo <- lm_robust(y ~ z, data = dat, fixed_effects = ~ x)
 
-test_that("tidy, summary, and print work", {
+test_that("tidy, glance, summary, and print work", {
 
   ## lm_robust
   lmo <- lm_robust(y ~ x, data = dat, se_type = "classical")
@@ -38,7 +38,6 @@ test_that("tidy, summary, and print work", {
     "data.frame"
   )
 
-
   expect_equal(
     nrow(tidy(lm_robust(y ~ x, data = dat, se_type = "classical"))),
     2
@@ -47,6 +46,17 @@ test_that("tidy, summary, and print work", {
   expect_equal(
     nrow(tidy(lmfo)),
     1
+  )
+
+  glance_lmo <- glance(lmo)
+
+  expect_is(glance_lmo , "data.frame")
+
+  expect_equal(nrow(glance_lmo), 1)
+
+  expect_equal(
+    colnames(glance(lmo)),
+    c("r.squared", "adj.r.squared", "statistic", "p.value", "df.residual", "N", "se_type")
   )
 
   expect_equal(
@@ -116,6 +126,11 @@ test_that("tidy, summary, and print work", {
     NA
   )
 
+  expect_error(
+    glance(lmrmo),
+    "Cannot use `glance` on linear models with multiple responses."
+  )
+
   lmroy <- lm_robust(y ~ z, data = dat, se_type = "classical")
   lmrox <- lm_robust(x ~ z, data = dat, se_type = "classical")
 
@@ -151,12 +166,20 @@ test_that("tidy, summary, and print work", {
     "data.frame"
   )
 
-
   capture_output(
     expect_equivalent(
       coef(summary(lmlo)),
       print(lmlo)
     )
+  )
+
+  glance_lmlo <- glance(lmlo)
+
+  expect_equal(nrow(glance_lmlo), 1)
+
+  expect_equal(
+    colnames(glance(lmlo)),
+    c("r.squared", "adj.r.squared", "statistic", "p.value", "df.residual", "N", "se_type")
   )
 
   ## horvitz_thompson
@@ -184,6 +207,15 @@ test_that("tidy, summary, and print work", {
     )
   )
 
+  glance_ht <- glance(ht)
+
+  expect_equal(nrow(glance_ht), 1)
+
+  expect_equal(
+    colnames(glance(ht)),
+    c("N", "se_type", "condition2", "condition1")
+  )
+
   ## difference_in_means
   dim <- difference_in_means(y ~ x, data = dat)
   expect_is(
@@ -201,6 +233,15 @@ test_that("tidy, summary, and print work", {
       coef(summary(dim)),
       print(dim)
     )
+  )
+
+  glance_dim <- glance(dim)
+
+  expect_equal(nrow(glance_dim), 1)
+
+  expect_equal(
+    colnames(glance(dim)),
+    c("design", "df", "N", "N_blocks", "N_clusters", "condition2", "condition1")
   )
 
   # rank deficient
@@ -375,9 +416,9 @@ test_that("coef and confint work", {
 
 test_that("predict works", {
   set.seed(42)
-  n <- 10
+  n <- 20
   dat <- data.frame(
-    x = rep(0:1, times = 5),
+    x = rep(0:1, times = 10),
     w = runif(n),
     z = rnorm(n),
     cl = as.factor(sample(letters[1:3], size = n, replace = T)),
@@ -404,9 +445,9 @@ test_that("predict works", {
   )
 
   # missingness
-  n <- 11
+  n <- 21
   new_dat <- data.frame(
-    x = rep(0:1, times = c(5, 6)),
+    x = rep(0:1, times = c(10, 11)),
     w = runif(n),
     z = rnorm(n),
     cl = as.factor(sample(letters[1:3], size = n, replace = T)),
@@ -487,9 +528,9 @@ test_that("predict works", {
   )
 
   # lm_lin
-  n <- 11
+  n <- 21
   new_dat <- data.frame(
-    x = rep(0:1, times = c(5, 6)),
+    x = rep(0:1, times = c(10, 11)),
     z = rnorm(n),
     cl = as.factor(sample(letters[1:3], size = n, replace = TRUE)),
     y = rnorm(n)